Seperated the function and variables to a different module.
Created a folder component and seperated function and variables in file months.js and style_vars.js correspondingly. Made necessary changes to App.js for importing and using them.
