+++ src/component/months.js
@@ -1,0 +1,8 @@
+let getMonthName = (number) => {
+    let monthArray = ['Jan', 'Feb', 'March', 'April', 'May', 'June', 'July', 'August', 'Sep', 'October', 'Nov', 'Dec'];
+
+    return monthArray[number];
+
+}
+
+export { getMonthName }
