+++ src/component/style_vars.js
@@ -1,0 +1,8 @@
+let header = {
+    backgroundColor: '#222'
+}
+let defaultTextColor = {
+    color: '#fff'
+} 
+
+export { header, defaultTextColor }
